backend_api/requirements.txt
fastapi
uvicorn
python-multipart
python-jose[cryptography]
passlib[bcrypt]
sqlalchemy
ultralytics
pillow
numpy